Assignment02_students/producers/requirements.txt
kafka-python
numpy
